[MEA-29] - Lista localuri
- Creat AbstractProfilePageController
