+++ src/main/java/controllers/ViewProfileAbstractController.java
@@ -1,13 +1,31 @@
 package controllers;
 
-import javafx.fxml.Initializable;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
+import models.EventCardModel;
 
-import java.net.URL;
-import java.util.ResourceBundle;
+public abstract class ViewProfileAbstractController extends AbstractProfilePageController {
+//    @FXML
+//    public TextField nameField;
+//
+//    @FXML
+//    public TextField emailField;
+//
+//    @FXML
+//    public TextField userTypeField;
 
-public class ViewProfileAbstractController implements Initializable {
-    @Override
-    public void initialize(URL location, ResourceBundle resources) {
+    @FXML
+    public TableView<EventModelContainer> eventsTableView;
 
-    }
+    @FXML
+    public TableColumn<EventModelContainer, EventCardModel> eventsTableColumn;
+
+    @FXML
+    public Button editProfilePageButton;
+
+    protected abstract void onEditProfilePageButtonClick(ActionEvent actionEvent);
 }
